+++ front/extension/app/src/lib/swr.ts
